+++ test/driver/run_as_ppx_rewriter/test.t/run.t
@@ -13,12 +13,15 @@ The registered rewriters get applied when using `run_as_ppx_rewriter` as entry p
   hi
   bye
 
-The driver's `shared_args` arguments should be taken into account with the
-exception of the perform check arguments.
+The driver's `shared_args` are taken into account, such as `-apply`...
 
-[To be fixed]
-But at the moment, it's the other way around: the check argument is taken into
-account...
+  $ ocaml -ppx 'print_greetings -apply print_hi' file.ml
+  hi
+  File "./file.ml", line 2, characters 11-20:
+  Error: Uninterpreted extension 'print_bye'.
+  [2]
+
+... and `-check`
 
   $ echo "[@@@attr non_registered_attr]" > attribute_file.ml
   $ ocaml -ppx 'print_greetings -check' attribute_file.ml
@@ -26,41 +29,63 @@ account...
   Error: Attribute `attr' was not used
   [2]
 
-[To be fixed]
-...while some of the other `shared_args` aren't
 
-  $ ocaml -ppx 'print_greetings -apply print_hi' file.ml
-  hi
-  bye
-
-[To be fixed]
-If a non-compatible file gets fed, the file name should get reported correctly,
-but isn't at the moment
+If a non-compatible file gets fed, the file name is reported correctly
 
   $ touch no_binary_ast.ml
   $ print_greetings no_binary_ast.ml some_output
-  End_of_file
-  [2]
+  File "no_binary_ast.ml", line 1:
+  Error: Expected a binary AST as input
+  [1]
 
 The only possible usage is [extra_args] <infile> <outfile>...
 
-  $ print_greetings some_input 2>&1 | sed 's/Usage: .*_build/Usage: (...)_build/'
-  Usage: (...)_build/default/test/driver/run_as_ppx_rewriter/print_greetings.exe [extra_args] <infile> <outfile>
+  $ print_greetings some_input
+  Usage: print_greetings.exe [extra_args] <infile> <outfile>
+  [2]
 
-[To be fixed]
 ...in particular the order between the flags and the input/output matters.
-That should get reported nicely, but isn't at the moment
 
   $ touch some_output
   $ print_greetings some_input some_output -check
-  End_of_file
+  print_greetings: anonymous arguments not accepted.
+  print_greetings.exe [extra_args] <infile> <outfile>
+    -loc-filename <string>      File name to use in locations
+    -reserve-namespace <string> Mark the given namespace as reserved
+    -no-check                   Disable checks (unsafe)
+    -check                      Enable checks
+    -no-check-on-extensions     Disable checks on extension point only
+    -check-on-extensions        Enable checks on extension point only
+    -no-locations-check         Disable locations check only
+    -locations-check            Enable locations check only
+    -apply <names>              Apply these transformations in order (comma-separated list)
+    -dont-apply <names>         Exclude these transformations
+    -no-merge                   Do not merge context free transformations (better for debugging rewriters)
+    -cookie NAME=EXPR           Set the cookie NAME to EXPR
+    --cookie                    Same as -cookie
+    -help                       Display this list of options
+    --help                      Display this list of options
   [2]
 
-[To be fixed]
-The only exception should be consulting help, but isn't at the moment
+The only exception is consulting help
 
-  $ print_greetings -help 2>&1 | sed 's/Usage: .*_build/Usage: (...)_build/'
-  Usage: (...)_build/default/test/driver/run_as_ppx_rewriter/print_greetings.exe [extra_args] <infile> <outfile>
+  $ print_greetings -help
+  print_greetings.exe [extra_args] <infile> <outfile>
+    -loc-filename <string>      File name to use in locations
+    -reserve-namespace <string> Mark the given namespace as reserved
+    -no-check                   Disable checks (unsafe)
+    -check                      Enable checks
+    -no-check-on-extensions     Disable checks on extension point only
+    -check-on-extensions        Enable checks on extension point only
+    -no-locations-check         Disable locations check only
+    -locations-check            Enable locations check only
+    -apply <names>              Apply these transformations in order (comma-separated list)
+    -dont-apply <names>         Exclude these transformations
+    -no-merge                   Do not merge context free transformations (better for debugging rewriters)
+    -cookie NAME=EXPR           Set the cookie NAME to EXPR
+    --cookie                    Same as -cookie
+    -help                       Display this list of options
+    --help                      Display this list of options
 
 [To be fixed]
 Binary AST's of any by ppxlib supported OCaml version should be supported as input;